Feature/macOS support based on async-ffi branch

+++ src/bindings/plugin_mgr.cc
@@ -91,6 +91,7 @@ nlohmann::json head::plugin_installer::install_plugin(const std::string& downloa
         m_updater->dispatch_progress("Starting Plugin Installer...", 0, false);
 
         const auto downloadPath = get_plugins_path();
+        std::filesystem::create_directories(downloadPath);
         std::string uuidStr = GenerateUUID();
         std::filesystem::path zipPath = downloadPath / uuidStr;
 
@@ -102,12 +103,15 @@ nlohmann::json head::plugin_installer::install_plugin(const std::string& downloa
 
         Http::DownloadWithProgress({ downloadUrl, totalSize }, zipPath, progressCallback);
         m_updater->dispatch_progress("Setting up installed plugin...", 50, false);
-        std::filesystem::create_directories(downloadPath);
-        Util::ExtractZipArchive(zipPath.string(), downloadPath.string(), [&](int current, int total, const char*)
+        const bool extracted = Util::ExtractZipArchive(zipPath.string(), downloadPath.string(), [&](int current, int total, const char*)
         {
             double percent = (double(current) / total) * 100.0;
             m_updater->dispatch_progress("Extracting plugin archive...", 50.0 + (45.0 * (percent / 100.0)), false);
         });
+
+        if (!extracted) {
+            throw std::runtime_error("Failed to extract plugin archive");
+        }
 
         m_updater->dispatch_progress("Cleaning up...", 95, false);
         std::filesystem::remove(zipPath);
